fee u32 -> f64

+++ json/src/lib.rs
@@ -1823,6 +1823,7 @@ pub struct DecodeRawTransactionResult {
 
 /// Model for Getblock tx verbosity 2
 #[derive(Clone, PartialEq, Eq, Debug, Deserialize, Serialize)]
+#[derive(Clone, PartialEq, Debug, Deserialize, Serialize)]
 pub struct GetBlockVerboseTransactionResult {
     pub txid: bitcoin::Txid,
     pub hash: bitcoin::Wtxid,
@@ -1833,7 +1834,7 @@ pub struct GetBlockVerboseTransactionResult {
     pub locktime: u32,
     pub vin: Vec<GetRawTransactionResultVin>,
     pub vout: Vec<GetRawTransactionResultVout>,
-    pub fee: Option<u32>,
+    pub fee: Option<f64>,
     #[serde(default, with = "crate::serde_hex")]
     pub hex: Vec<u8>,
 }
